+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file. See [Keep a
 
 ## Unreleased
 
+### Added
+
+- Added a `--version` flag. The version must be defined when compiling the interpreter, with `-DVERSION=...` (gcc/clang) or `/DVERSION=...` (MSVC).
+
 ## [1.3.5] - 2023-06-11
 
 No code changes, just fixing CI.